TweetIt/requirements.txt
Django==2.2
django-crispy-forms==1.7.2
djangorestframework==3.9.2
Pillow==5.4.1
pytz==2018.9
sqlparse==0.3.0
